econ-agent/requirements.txt
openai>=1.42.0
feedparser>=6.0.11
python-dotenv>=1.0.1
yfinance>=0.2.52
pandas>=2.2.2
rich>=13.7.1
tzdata>=2024.1
